Add Compose Map & Compose Set
Compose Map takes an array of keys and an array of values and composes
a Map from it.

Compose Set takes an array of elements and composes a Set from it.

There are no unit tests for these functions because this project is
maintained in LabVIEW 2017 and Maps and Sets were added in LabVIEW 2019.