Fix error 1004 when (de-)composing interfaces
This issue only appears in LabVIEW version 2020 and later. When a class
implements an interface, the interface is listed as ancestor.

Error 1004 happens when trying to load the private data control for all
ancestors in the hierarchy, because interfaces don't have private data
controls.

Fixed by checking if the private data control exists before calling
Open VI Reference on it.

Fixes #6