Fix error 1401 in Compose LabVIEW Object
This error happens when the version number of a child class is higher
than the version number of its parent.

The reason for this is that the version numbers are stored in reverse
order, which is not implemented properly by Compose LabVIEW Object. It
assembles the object with version numbers in the wrong order, which can
work fine if mutation history is maintained (there are probably some
cases where this still fails).

Fixed by storing version numbers in the correct order.

Closes #7