Fix error 1013 when accessing panel for class in executable
This error only occurs in the runtime environment and only for some
configurations. In LV2017 the issue does not occur when 'Use fast file
format' is checked.

Error 1013 indicates that the front panel is not compiled into the
executable. Using the 'Connector Pane:Data Type' property[1] fixes
this issue. However, this property only returns the data type and not
the default value, which is why other tests, that rely on the default
value of an object, now fail.

[1] https://zone.ni.com/reference/en-XX/help/371361R-01/lvscript/vi_connector_pane058datatype/

References #12