Rebuild object lookup code
The current lookup code goes through each class in the current instance
and checks if it is a parent of the current class, recursively. Then it
traverses the object hierarchy to retrieve version and type information
for the private data control, which is accessed by opening a reference
to it.

This process is not ideal and requires a lot of error handling to make
it work. There is a huge performance impact, which is why caching was
implemented.

60859f9 introduced a change which no longer returns the default data of
an object but only its type. While this change may not be an issue for
most users, it does limit the functionality of this library - especially
for users who rely on the default value of an object (i.e. when filling
in blanks).

The solution in this commit replaces the existing lookup code by a much
simpler solution using the >hidden< 'Get Mutation History.vi', which is
part of the LabVIEW installation.

Instead of testing each class in the application instance, it can simply
get the parent class name and walk straight up. The returned data also
includes the class default value, which is the type and data of the
private data control. This simplifies the entire process by a lot and
should(tm) run much more efficient than the previous solution.

Disclaimer: I have not done any performance benchmarks. The main focus
of this commit is feature completeness - so full access to the default
values of objects.