Fix error 1400 when composing interfaces
Compose LabVIEW Object returns error 1400 when composing interfaces
that were previously decomposed with Decompose LabVIEW Object. This
happens because the values for `Private Data Clusters` and `Library
Versions` are empty.

For interfaces (and instances of LabVIEW Object), there is no need
to compose the object at all, because there is nothing to compose.

Fixed by turning Compose LabVIEW Object into a no-op for interfaces
and instances of LabVIEW Object.

Closes #14