Skip object composition for empty arrays
When the version info and private data control arrays are empty, it is 
safe to assume that nothing useful can be constructed from it. In the 
case of interfaces, both arrays are always empty, which means that 
skipping object composition effectively enables interface support by 
passing through the original type.