Replace interface detection logic
The current approach of looking up an object's private data control in 
the list of VIs in memory does not work reliably because of a bug in 
LabVIEW that results in the `AllVIs` property to return an incomplete 
set of VIs for dynamically loaded classes when the user performs other 
actions (e.g., selecting a FP or BD).

This is a new approach, using a constant of an interface's private data 
control from LV2025Q3 (saved back for LV2017) to compare against the 
current private data control. It should work reliably because the 
private data control of an interface is reported as a cluster with zero 
elements, which cannot exist for a regular class because it would be 
broken.